Add ConsoleFiltersTimePeriodCell
