+++ Sources/PulseUI/Features/Filters/ConsoleSharedFiltersView.swift
@@ -14,18 +14,10 @@ struct ConsoleSharedFiltersView: View {
     }
 
     private var sectionTimePeriod: some View {
-#if os(iOS)
-        Section(
-            content: { timePeriodContent },
-            header: { timePeriodHeader },
-            footer: { quickFilters }
-        )
-#else
         ConsoleFilterSection(
             header: { timePeriodHeader },
-            content: { timePeriodContent }
+            content: { ConsoleFiltersTimePeriodCell(selection: $viewModel.criteria.dates) }
         )
-#endif
     }
 
     private var timePeriodHeader: some View {
@@ -36,42 +28,6 @@ struct ConsoleSharedFiltersView: View {
             isDefault: viewModel.isDatesDefault,
             isEnabled: $viewModel.criteria.dates.isEnabled
         )
-    }
-
-    @ViewBuilder
-    private var timePeriodContent: some View {
-#if os(iOS) || os(macOS)
-        DateRangePicker(title: "Start", date: $viewModel.criteria.dates.startDate)
-        DateRangePicker(title: "End", date: $viewModel.criteria.dates.endDate)
-#endif
-
-#if os(tvOS) || os(watchOS)
-        Picker("Date Range", selection: $viewModel.criteria.quickDatesFilter) {
-            ForEach(ConsoleDatesQuickFilter.allCases, id: \.self) {
-                Text($0.title).tag($0)
-            }
-        }
-#endif
-
-#if os(macOS)
-        quickFilters
-#endif
-    }
-
-    @ViewBuilder
-    private var quickFilters: some View {
-        HStack(alignment: .center, spacing: 10) {
-            Text("Quick Filters")
-                .lineLimit(1)
-                .foregroundColor(.secondary)
-                .frame(maxWidth: .infinity, alignment: .leading)
-            Button("Session") { viewModel.criteria.dates = .session }
-            Button("Recent") { viewModel.criteria.dates = .recent }
-            Button("Today") { viewModel.criteria.dates = .today }
-        }
-        .buttonStyle(.plain)
-        .padding(.top, 4)
-        .foregroundColor(.accentColor)
     }
 
     // MARK: Filters
