+++ Sources/PulseUI/Features/Filters/SearchCriteria/ConsoleMessageSearchCriteria+Predicate.swift
@@ -27,17 +27,6 @@ extension ConsoleMessageSearchCriteria {
             }
         }
 
-#if os(tvOS) || os(watchOS)
-        if let dates = shared.quickDatesFilter.makeDateFilter() {
-            if let startDate = dates.startDate {
-                predicates.append(NSPredicate(format: "createdAt >= %@", startDate as NSDate))
-            }
-            if let endDate = shared.dates.endDate {
-                predicates.append(NSPredicate(format: "createdAt <= %@", endDate as NSDate))
-            }
-        }
-#endif
-
         if shared.filters.isEnabled {
             if shared.filters.inOnlyPins {
                 predicates.append(NSPredicate(format: "isPinned == YES"))
